static/functions/upload/requirements.txt
orator
PyMySQL
PyJWT
boto3
